requirements.txt
Django==1.4.1
South==0.7.6
django-crispy-forms==1.1.4
django-olwidget==0.50
psycopg2==2.4.5
